Split out the the node details tab
