+++ src/main/java/net/sourceforge/pmd/util/fxdesigner/app/NodeSelectionSource.java
@@ -9,6 +9,7 @@ import static java.util.Collections.emptySet;
 import java.util.Set;
 
 import org.reactfx.EventStream;
+import org.reactfx.value.Val;
 
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.util.fxdesigner.XPathPanelController;
@@ -46,12 +47,14 @@ public interface NodeSelectionSource extends ApplicationComponent {
      *                              For now some must, because they aggregate several selection sources (the {@link net.sourceforge.pmd.util.fxdesigner.NodeInfoPanelController}).
      *                              Splitting it into separate controls will remove the need for that.
      */
-    default void initNodeSelectionHandling(DesignerRoot root,
-                                           EventStream<? extends NodeSelectionEvent> mySelectionEvents,
-                                           boolean alwaysHandleSelection) {
+    default Val<Node> initNodeSelectionHandling(DesignerRoot root,
+                                                EventStream<? extends NodeSelectionEvent> mySelectionEvents,
+                                                boolean alwaysHandleSelection) {
         MessageChannel<NodeSelectionEvent> channel = root.getService(DesignerRoot.NODE_SELECTION_CHANNEL);
         mySelectionEvents.subscribe(n -> channel.pushEvent(this, n));
-        channel.messageStream(alwaysHandleSelection, this).subscribe(evt -> setFocusNode(evt.selected, evt.options));
+        Val<NodeSelectionEvent> selection = channel.latestMessage(alwaysHandleSelection, this);
+        selection.values().subscribe(evt -> setFocusNode(evt.selected, evt.options));
+        return selection.map(it -> it.selected);
     }
